fold AbstractCrypticTicketRegistry up to AbstractDistributedTicketRegistry
